update e2fsprogs
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@33175 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: ccff1abef9e5a0234623d2b5f5f41b84b714d1d9

+++ src/router/e2fsprogs/lib/ext2fs/bitops.c
@@ -138,7 +138,7 @@ unsigned int ext2fs_bitcount(const void *addr, unsigned int nbytes)
 	const __u32 *p;
 	unsigned int res = 0;
 
-	while (((((unsigned long) cp) & 3) != 0) && (nbytes > 0)) {
+	while (((((uintptr_t) cp) & 3) != 0) && (nbytes > 0)) {
 		res += popcount8(*cp++);
 		nbytes--;
 	}
